pinecone-comments/requirements.txt
transformers[torch]==4.1.1
sentence-transformers
openai
langchain
pandas
numpy
pinecone-client
tiktoken
tqdm